# Alt.NET-Swagger/src/SwaggerDemo/APIClient/AltNetClient/Models/CompanyDTO.cs
// IMPORTANT: This class is auto-generated. Any changes to it will be lost
// during proxy re-generation. Generated by AutoRest from
// http://swagger.localtest.me/swagger/docs/v1. Visit
// https://github.com/Azure/autorest for details and usage examples.

namespace APIClient.AltNetClient.Models
{
    using System;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    /// <summary>
    /// DTO to describe company, it's employees and active projects
    /// </summary>
    public partial class CompanyDTO
    {
        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "Id")]
        public string Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "CompanyName")]
        public string CompanyName { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "Location")]
        public string Location { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "StockTicker")]
        public string StockTicker { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "Employees")]
        public IList<UserDTO> Employees { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "ActiveProjects")]
        public IList<ProjectDTO> ActiveProjects { get; set; }

    }
}