# Alt.NET-Swagger/src/SwaggerDemo/APIClient/AltNetClient/Models/ProjectDTO.cs
// IMPORTANT: This class is auto-generated. Any changes to it will be lost
// during proxy re-generation. Generated by AutoRest from
// http://swagger.localtest.me/swagger/docs/v1. Visit
// https://github.com/Azure/autorest for details and usage examples.

namespace APIClient.AltNetClient.Models
{
    using System;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    /// <summary>
    /// </summary>
    public partial class ProjectDTO
    {
        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "Id")]
        public string Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "ProjectName")]
        public string ProjectName { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "DevEnvironmentUrl")]
        public string DevEnvironmentUrl { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "ProductionEnvironmentUrl")]
        public string ProductionEnvironmentUrl { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "MainDevelopers")]
        public IList<UserDTO> MainDevelopers { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "Testers")]
        public IList<UserDTO> Testers { get; set; }

    }
}