# Alt.NET-Swagger/src/SwaggerDemo/APIClient/AltNetClient/Models/UserDTO.cs
// IMPORTANT: This class is auto-generated. Any changes to it will be lost
// during proxy re-generation. Generated by AutoRest from
// http://swagger.localtest.me/swagger/docs/v1. Visit
// https://github.com/Azure/autorest for details and usage examples.

namespace APIClient.AltNetClient.Models
{
    using System;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    /// <summary>
    /// User data structure to return from API
    /// </summary>
    public partial class UserDTO
    {
        /// <summary>
        /// User Id
        /// </summary>
        [JsonProperty(PropertyName = "Id")]
        public int? Id { get; set; }

        /// <summary>
        /// Full name
        /// </summary>
        [JsonProperty(PropertyName = "Name")]
        public string Name { get; set; }

        /// <summary>
        /// User's Secret tag
        /// </summary>
        [JsonProperty(PropertyName = "Tag")]
        public string Tag { get; set; }

        /// <summary>
        /// Current Employer
        /// </summary>
        [JsonProperty(PropertyName = "Company")]
        public string Company { get; set; }

        /// <summary>
        /// Preferred programming language
        /// </summary>
        [JsonProperty(PropertyName = "PreferredLanguage")]
        public string PreferredLanguage { get; set; }

    }
}