# Alt.NET-Swagger/src/SwaggerDemo/AltNetClient.swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Alt.NET Demo ","description":"Swagger API Demo for ALT.NET Melbourne","contact":{"url":"http://swagger.io"}},"host":"swagger.localtest.me:80","schemes":["http"],"paths":{"/api/company/{id}":{"get":{"tags":["Company"],"operationId":"Company_GetCompany","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"companyId","in":"query","required":true,"type":"string"},{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CompanyDTO"}}},"deprecated":false},"put":{"tags":["Company"],"operationId":"Company_UpdateCompany","consumes":["application/json","text/json","application/x-www-form-urlencoded"],"produces":["application/json","text/json"],"parameters":[{"name":"companyId","in":"query","required":true,"type":"string"},{"name":"updatedData","in":"body","required":true,"schema":{"$ref":"#/definitions/CompanyDTO"}},{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CompanyDTO"}}},"deprecated":false},"delete":{"tags":["Company"],"operationId":"Company_DeleteCompany","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"companyId","in":"query","required":true,"type":"string"},{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Object"}}},"deprecated":false}},"/api/company":{"get":{"tags":["Company"],"operationId":"Company_GetList","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"pageSize","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/CompanyDTO"}}}},"deprecated":false},"post":{"tags":["Company"],"operationId":"Company_CreateCompany","consumes":["application/json","text/json","application/x-www-form-urlencoded"],"produces":["application/json","text/json"],"parameters":[{"name":"data","in":"body","required":true,"schema":{"$ref":"#/definitions/CompanyDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Object"}}},"deprecated":false}},"/api/company/{companyId}/projects":{"get":{"tags":["Company"],"operationId":"Company_GetCompanyProjects","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"companyId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ProjectDTO"}}}},"deprecated":false}},"/api/project/{projectId}":{"get":{"tags":["Project"],"operationId":"Project_GetProjectById","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProjectDTO"}}},"deprecated":false}},"/api/project/{projectId}/users":{"get":{"tags":["Project"],"operationId":"Project_GetProjectDevelopers","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}}}},"deprecated":false}},"/api/user/{id}":{"get":{"tags":["User"],"summary":"Get specific user by Id","description":"This operation is expensive to call as it'll lookup in multiple databases.","operationId":"User_GetUserById","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"id","in":"path","description":"Id of the user to find.","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"User found.","schema":{"$ref":"#/definitions/UserDTO"}},"400":{"description":"Id parameter has invalid format or out of expected range."},"404":{"description":"User with passed Id doesn't exists."}},"deprecated":false},"put":{"tags":["User"],"summary":"Update user. User Id must be &gt; 10","operationId":"User_UpdateUser","consumes":["application/json","text/json","application/x-www-form-urlencoded"],"produces":["application/json","text/json"],"parameters":[{"name":"id","in":"path","description":"User Id to update","required":true,"type":"integer","format":"int32"},{"name":"user","in":"body","description":"Updated user model.","required":true,"schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"201":{"description":"User was updated in the database","schema":{"$ref":"#/definitions/Object"}},"400":{"description":"Incorrect data format"},"500":{"description":"We are sorry, something went wrong."}},"deprecated":false},"delete":{"tags":["User"],"summary":"Delete user. User Id must be &gt; 10","operationId":"User_DeletUserById","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"id","in":"path","description":"Id of the user to delete","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"User was deleted.","schema":{"$ref":"#/definitions/Object"}},"400":{"description":"Bad request."},"500":{"description":"We are sorry, something went wrong."}},"deprecated":false}},"/api/user":{"post":{"tags":["User"],"summary":"Create a new user, user Id must be &gt; 10","operationId":"User_CreateUser","consumes":["application/json","text/json","application/x-www-form-urlencoded"],"produces":["application/json","text/json"],"parameters":[{"name":"user","in":"body","description":"User model.","required":true,"schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"201":{"description":"User was added to the database","schema":{"$ref":"#/definitions/Object"}},"400":{"description":"Incorrect data format"},"409":{"description":"User with this Id already exists"},"500":{"description":"We are sorry, something went wrong."}},"deprecated":false}},"/api/user/list":{"get":{"tags":["User"],"summary":"List all users available for contracts","operationId":"User_ListAllUsers","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"pageSize","in":"query","description":"Number of records to return","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}}}},"deprecated":false}},"/api/user/list/byproject/{projectId}":{"get":{"tags":["User"],"summary":"List all users for the specific project","operationId":"User_ListAllUsersByProject","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"projectId","in":"path","description":"Project ID","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}}}},"deprecated":false}}},"definitions":{"CompanyDTO":{"description":"DTO to describe company, it's employees and active projects","type":"object","properties":{"Id":{"type":"string"},"CompanyName":{"type":"string"},"Location":{"type":"string"},"StockTicker":{"type":"string"},"Employees":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}},"ActiveProjects":{"type":"array","items":{"$ref":"#/definitions/ProjectDTO"}}}},"UserDTO":{"description":"User data structure to return from API","type":"object","properties":{"Id":{"format":"int32","description":"User Id","type":"integer"},"Name":{"description":"Full name","type":"string"},"Tag":{"description":"User's Secret tag","type":"string"},"Company":{"description":"Current Employer","type":"string"},"PreferredLanguage":{"description":"Preferred programming language","type":"string"}}},"ProjectDTO":{"type":"object","properties":{"Id":{"type":"string"},"ProjectName":{"type":"string"},"DevEnvironmentUrl":{"type":"string"},"ProductionEnvironmentUrl":{"type":"string"},"MainDevelopers":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}},"Testers":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}}}},"Object":{"type":"object","properties":{}}}}